Tap Flow: Revise callbacks for easier use.
Fixes https://github.com/getreuer/qmk-modules/issues/2

Thanks to @JJGadgets for feedback about this.

Currently, it is possible but cumbersome to change the set of keys where
Tap Flow is enabled. It requires the user to define multiple functions
in keymap.c. This commit revises to introduce a simpler callback
`is_tap_flow_key()` to make such use easier to specify. The default
definition of this callback corresponds to

    bool is_tap_flow_key(uint16_t keycode) {
      switch (keycode) {
        case QK_MOD_TAP ... QK_MOD_TAP_MAX:
          keycode = QK_MOD_TAP_GET_TAP_KEYCODE(keycode);
          break;
        case QK_LAYER_TAP ... QK_LAYER_TAP_MAX:
          keycode = QK_LAYER_TAP_GET_TAP_KEYCODE(keycode);
          break;
      }
      switch (keycode) {
        case KC_SPC:
        case KC_A ... KC_Z:
        case KC_DOT:
        case KC_COMM:
        case KC_SCLN:
        case KC_SLSH:
          return true;
      }
      return false;
    }

The former `get_tap_flow()` callback has been renamed to
`get_tap_flow_term()`, and can still be used for fine-grained
configuration, specifying different per-key timeouts and conditions
based on the previous key.

See the documentation for details:
https://getreuer.info/posts/keyboards/tap-flow/index.html