Speculative Hold: Apply mod-tap mods immediately.
Speculative Hold is a new community module that applies mod-tap mods
immediately on key down. Then depending on how the mod-tap key is
settled, the hold action is canceled (when the mod-tap is being tapped)
or allowed to continue (when the mod-tap is being held).

"Speculative" holding behavior is useful to remove the input lag that
mod-tap keys normally have. An acute case is when using a mod-tap with
an external mouse to perform Ctrl + click, as this would normally
require waiting out the tapping term. With Speculative Hold, the mod is
activated immediately.

For full documentation, see
<https://getreuer.info/posts/keyboards/speculative-hold>