Lumino: lumino_wake() tells lighting to wake up.
This commit adds function `lumino_wake(event_count)`, used to indicate
or simulate user activity. It tells the lighting to wake up (if
sleeping) and adds `event_count` to the activity counter. The argument
is typically `event_count = 1` to consider the call as representing one
key event's worth of activity.